добавляет сброс полей формы и возврат к начальному состоянию
