+++ js/util.js
@@ -1,5 +1,3 @@
-const ALERT_POPUP_TIME = 5000;
-
 const getRandomBetween = (min, max) => {
   let roundedMin = Math.ceil(min);
   let roundedMax = Math.floor(max);
@@ -79,25 +77,6 @@ const pluralize = (count, variants) => {
   return variants[2];
 };
 
-const showAlertPopup = (message) => {
-  const alertContainer = document.createElement('div');
-  alertContainer.style.zIndex = 1000;
-  alertContainer.style.position = 'fixed';
-  alertContainer.style.left = 0;
-  alertContainer.style.top = 0;
-  alertContainer.style.right = 0;
-  alertContainer.style.padding = '10px 3px';
-  alertContainer.style.fontFamily = 'Roboto';
-  alertContainer.style.fontSize = '30px';
-  alertContainer.style.textAlign = 'center';
-  alertContainer.style.backgroundColor = 'red';
-  alertContainer.textContent = message;
-  document.body.append(alertContainer);
-  setTimeout(() => {
-    alertContainer.remove();
-  }, ALERT_POPUP_TIME);
-}
-
 export {
   getRandomBetween,
   getRandomFloat,
@@ -105,6 +84,4 @@ export {
   getRandomArrayWithUniqueElements,
   getRandomObjectValue,
   pluralize,
-  showAlertPopup,
-  ALERT_POPUP_TIME,
 };
